% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xval.R
\name{a4ahcxval}
\alias{a4ahcxval}
\title{Compute a retrospective hindcast cross-validation of a4a stock and indices}
\usage{
a4ahcxval(stock, indices, nyears = 5, nsq = 3, fixed.ks = FALSE, ...)
}
\arguments{
\item{stock}{Input FLStock object.}

\item{indices}{Input FLIndices object.}

\item{nyears}{Number if years for retrospective, defaults to 5.}

\item{nsq}{Number of years for average biology and selectivity, defaults to 3.}

\item{fixed.ks}{Is the number of knots is splines with 'year' constant?}

\item{...}{Any submodels and other arguments for the call to \emph{sca}.}
}
\value{
A list containing elements 'stock', of class \emph{FLStocks}, and
'indices', a list of \emph{FLIndices} objects.
}
\description{
Compute a retrospective hindcast cross-validation of a4a stock and indices
}
