% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a4aLAK.R
\name{a4aLAK}
\alias{a4aLAK}
\title{Convert length frequency to age composition data}
\usage{
a4aLAK(
  Ldat,
  linf = 100,
  k = 0.2,
  t0 = -0.5,
  cvL = 0.15,
  ages = 0:5,
  sd.fix = FALSE,
  plot = FALSE,
  aW = NULL,
  bW = NULL
)
}
\arguments{
\item{Ldat}{a matrix with two columns, where each row specifies the lower and upper length for a given length bin (the lowest should be -Inf, and highest Inf)}

\item{linf}{Infinitve length}

\item{k}{Brody growth coefficient}

\item{t0}{theoretical age at zero Length}

\item{cvL, }{CV of mean length-at-age}

\item{ages}{desired age range}

\item{sd.fix}{if TRUE a fix sd at length-at-age is assumed corresponding to 0.5*Linf}

\item{aW}{parameter of length-weight in mm}

\item{bW}{parameter of length-weight relationship}
}
\value{
AgeComp_at, a matrix of expected age-composition data, where columns are samples in year t, and cells are the count of samples with a given age and year
}
\description{
Converts a data.frame of multi-annual length frequency into age composition data
}
