% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a4aUtils.R
\name{a4apar}
\alias{a4apar}
\title{a4apar()}
\usage{
a4apar(
  mfrow = c(1, 1),
  plot.cex = 1,
  mai = c(0.55, 0.6, 0.1, 0.1),
  omi = c(0, 0, 0, 0) + 0.1,
  labs = TRUE
)
}
\arguments{
\item{mfrow}{determines plot frame set up}

\item{plot.cex}{cex graphic option}

\item{mai}{graphical par for plot margins}

\item{labs}{if TRUE margins are narrow}
}
\description{
Set the par() to options suitable for ss3diags multi plots
}
