% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sol274}
\alias{sol274}
\alias{stock}
\alias{indices}
\alias{fit}
\title{Data from the 2020 ICES stock assessment of North Sea sole (sol.27.4)}
\format{
Objects of class FLStock, FLIndices and a4aFitSA
}
\description{
fmod <- ~te(replace(age, age > 8, 8), year, k = c(4, 22)) +
s(replace(age, age > 8, 8), k=4) +
s(year, k=22, by=as.numeric(age==1))
}
\details{
qmod <- list(~s(age, k=3), ~s(age, k=3))

vmod <- list(
~s(age, k=3),
~s(age, k=3),
~s(age, k=3))

srmod <- ~factor(year)
}
\keyword{datasets}
