% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a4afits.r
\name{plota4aBioidx}
\alias{plota4aBioidx}
\title{plota4aBioidx {{{
Computes observed and expected FLIndexBiomass from a4a fits}
\usage{
plota4aBioidx(stock, fit, indices)
}
\arguments{
\item{stock}{Input FLStock object.}

\item{fit}{object a4a fit .}

\item{indices}{Input FLIndices object.}

\item{nyears}{Number if years for retrospective, defaults to 5.}

\item{nsq}{Number of years for average biology and selectivity, defaults to 3.}

\item{fixed.ks}{Is the number of knots is splines with 'year' constant?}
}
\value{
ggplot
}
\description{
plota4aBioidx {{{
Computes observed and expected FLIndexBiomass from a4a fits
}
\examples{
data(sol274)
 # models
fmod <- ~te(replace(age, age > 8, 8), year, k = c(4, 22)) +
  s(replace(age, age > 8, 8), k=4) +
  s(year, k=22, by=as.numeric(age==1))
qmod <- list(~s(age, k=3), ~s(age, k=3))
vmod <- list(~s(age, k=3), ~s(age, k=3), ~s(age, k=3))
srmod <- ~factor(year)
# RUN a4a
fit <- sca(stock, indices, fmodel=fmod, qmodel=qmod, vmodel=vmod,sr=srmod)
idxs = a4aBioidx(stock,fit,indices)
plota4aBioidx(stock,fit,indices)
}
