% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runstest.R
\name{plotRunstest,a4aFitSA,FLIndices-method}
\alias{plotRunstest,a4aFitSA,FLIndices-method}
\title{Plot the runs test result for one or more time series}
\usage{
\S4method{plotRunstest}{a4aFitSA,FLIndices}(fit, obs, combine = TRUE)
}
\arguments{
\item{fit}{The result of a model fit.}

\item{obs}{The observations used in the fit.}

\item{combine}{Should ages be combined by addition, defaults to TRUE.}

\item{...}{Extra arguments.}
}
\value{
An object of class ggplot2::gg
}
\description{
Plot the runs test result for one or more time series
}
\examples{
data(sol274)
plotRunstest(fit, indices)
}
