% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runstest.R
\name{plotRunsTest}
\alias{plotRunsTest}
\alias{plotRunsTest,FLQuants,FLQuants-method}
\alias{plotRunsTest,FLQuant,FLQuant-method}
\alias{plotRunsTest,a4aFitSA,FLIndices-method}
\title{Plot the runs test result for one or more time series}
\usage{
plotRunsTest(fit, obs, ...)

\S4method{plotRunsTest}{FLQuants,FLQuants}(fit, obs, combine = TRUE)

\S4method{plotRunsTest}{FLQuant,FLQuant}(fit, obs, combine = TRUE)

\S4method{plotRunsTest}{a4aFitSA,FLIndices}(fit, obs, combine = TRUE)
}
\arguments{
\item{fit}{The result of a model fit.}

\item{obs}{The observations used in the fit.}

\item{...}{Extra arguments.}

\item{combine}{Should ages be combined by addition, defaults to TRUE.}
}
\value{
An object of class ggplot2::gg
}
\description{
Plot the runs test result for one or more time series
}
\examples{
data(sol274)
}
