% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xval.R
\name{plotXval}
\alias{plotXval}
\title{Plot of FLIndices cross-validation by retrospective hindcast}
\usage{
plotXval(x, y, order = "inverse")
}
\arguments{
\item{x}{An \emph{FLIndices} object of the original observations.}

\item{y}{A list contaning \emph{FLIndices} objects returned by \emph{xval}.}

\item{order}{Order in which retrospective runs are stored, defaults to"inverse".}
}
\value{
A ggplot object
}
\description{
Plot of FLIndices cross-validation by retrospective hindcast
}
\examples{
# SEE vignette
}
