% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runstest.R
\name{runstest,a4aFitSA,FLIndices-method}
\alias{runstest,a4aFitSA,FLIndices-method}
\title{Computes Runs Test p-values}
\usage{
\S4method{runstest}{a4aFitSA,FLIndices}(fit, obs, combine = TRUE)
}
\arguments{
\item{fit}{The result of a model fit.}

\item{obs}{The observations used in the fit.}

\item{combine}{Should ages be combined by addition, defaults to TRUE.}

\item{...}{Extra arguments.}
}
\value{
A list with elements 'p.values' and 'pass'.
}
\description{
Computes Runs Test p-values
}
\examples{
data(sol274)
# Call on a a4aFitSA object
runstest(sol274fit, indices)
#
runstest(index(sol274fit), lapply(indices, index))
}
