% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runstest.R
\name{runstest}
\alias{runstest}
\alias{runstest,FLQuant,FLQuant-method}
\title{Computes Runs Test p-values}
\usage{
runstest(fit, obs, ...)

\S4method{runstest}{FLQuant,FLQuant}(fit, obs, combine = TRUE)
}
\arguments{
\item{fit}{The result of a model fit.}

\item{obs}{The observations used in the fit.}

\item{...}{Extra arguments.}

\item{combine}{Should ages be combined by addition, defaults to TRUE.}
}
\value{
A list with runs 'p.values' and 'pass'
}
\description{
Computes Runs Test p-values
}
\examples{
data(sol274)
# Call on a a4aFitSA object
runstest(fit, indices)
#
runstest(index(fit), lapply(indices, index))
}
