% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xval.R
\name{xval}
\alias{xval}
\title{Compute a retrospective hindcast cross-validation of a4a stock and indices}
\usage{
xval(stock, indices, nyears = 5, nsq = 3, fixed.ks = FALSE, ...)
}
\arguments{
\item{stock}{Input FLStock object.}

\item{indices}{Input FLIndices object.}

\item{...}{Any arguments for the call to \emph{sca}.}

\item{nyears=5}{Number iof years for retrospective.}

\item{nsq=3}{Number of years for average biology and selectivity.}
}
\value{
A list containing elements 'stock', of class \emph{FLStocks}, and
'indices', a list of \emph{FLIndices} objects.
}
\description{
Compute a retrospective hindcast cross-validation of a4a stock and indices
}
