% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xval.R
\name{a4ahcxval}
\alias{a4ahcxval}
\title{Compute a retrospective hindcast cross-validation of a4a stock and indices}
\usage{
a4ahcxval(stock, indices, nyears = 5, nsq = 3, check.ks = FALSE, ...)
}
\arguments{
\item{stock}{Input FLStock object.}

\item{indices}{Input FLIndices object.}

\item{nyears}{Number if years for retrospective, defaults to 5.}

\item{nsq}{Number of years for average biology and selectivity, defaults to 3.}

\item{...}{Any submodels and other arguments for the call to \emph{sca}.}

\item{fixed.ks}{Is the number of knots is splines with 'year' constant?}
}
\value{
A list containing elements 'stocks', of class \emph{FLStocks}, and
'indices', a list of \emph{FLIndices} objects. See details for structure of this list.
}
\description{
The output of \code{a4ahcxval} consist of a list with two elements, named 'stocks'
and 'indices'. The first is an object of class \code{FLStocks}, each a peel from
the restrospective run. The second element is a list of \code{FLIndices} object.
The first \code{FLIndices} object, named 'data', is a copy of the input 'indices'
argument, with the additoned \code{catch.n} slot, if originally missing. The next
element, named as the final year of the data set, contains the naive prediction
of the input \code{FLIndices}, while the remaining elements are the result of a
hindcast prediction of the relevant indices, those within the year range of
as set ny \code{nyears}.
}
\examples{
data(sol274)
 # models
fmod <- ~te(replace(age, age > 8, 8), year, k = c(4, 22)) +
  s(replace(age, age > 8, 8), k=4) +
  s(year, k=22, by=as.numeric(age==1))
qmod <- list(~s(age, k=3), ~s(age, k=3))
vmod <- list(~s(age, k=3), ~s(age, k=3), ~s(age, k=3))
srmod <- ~factor(year)
# RUN xval
xval <- a4ahcxval(stock, indices, fmodel=fmod, qmodel=qmod, vmodel=vmod)
# PLOT result
plotXval(xval$indices)
}
