% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sol274}
\alias{sol274}
\alias{stock}
\alias{indices}
\alias{fit}
\title{Data from the 2020 ICES stock assessment of North Sea sole (sol.27.4)}
\format{
Objects of class FLStock, FLIndices and a4aFitSA
}
\description{
Aliquam sagittis feugiat felis eget consequat. Praesent eleifend dolor massa,
vitae faucibus justo lacinia a. Cras sed erat et magna pharetra bibendum quis in
mi. Sed sodales mollis arcu, sit amet venenatis lorem fringilla vel. Vivamus vitae
ipsum sem. Donec malesuada purus at libero bibendum accumsan. Donec ipsum sapien,
feugiat blandit arcu in, dapibus dictum felis.
}
\details{
fmod <- ~te(replace(age, age > 8, 8), year, k = c(4, 22)) +
s(replace(age, age > 8, 8), k=4) +
s(year, k=22, by=as.numeric(age==1))

qmod <- list(~s(age, k=3), ~s(age, k=3))

vmod <- list(
~s(age, k=3),
~s(age, k=3),
~s(age, k=3))

srmod <- ~factor(year)
}
\keyword{datasets}
